Dijkstras.c
Since Dijkstra follows a Greedy Approach, once a node is marked as visited it will not be revisited even if there is another path with less cost or distance. This is an issue when there is a negative edge in the graph. Hence,Dijkstras algo will not work for a directed graph with negative edges. If we intend to find the shortest path of a graph with negative edges, we use bellman ford algorithm